* commit 'beb96e7eeffcb2964f9bba9b1428d60befedaf37':
  Fixed BuildCheckedDiv signature

+++ compiler/optimizing/builder.cc
@@ -547,7 +547,7 @@ bool HGraphBuilder::BuildStaticFieldAccess(const Instruction& instruction,
 
 void HGraphBuilder::BuildCheckedDiv(uint16_t out_reg,
                                     uint16_t first_reg,
-                                    uint16_t second_reg,
+                                    int32_t second_reg,
                                     uint32_t dex_offset,
                                     Primitive::Type type,
                                     bool second_is_lit) {
